refactor(server): /api/setup 라우트 분할 (Plan C #C9 7단계)
- src/server/routes/setup.ts 신규: 5개 라우트 캡슐화
  (validate-token, repos, labels, preview, apply)
- generateSetupYaml + computeYamlDiff 헬퍼 + Setup*BodySchema 인라인 스키마 이전
- safeError 적용
- dashboard-api.ts에서 unused import 정리 (readFileSync/writeFileSync/copyFileSync/
  mkdirSync/homedir)
- dashboard-api.ts 1268 → 1058줄 (-210)

남은 도메인: repositories/health/doctor/sse/skip-events/auth/new-issue
